<Feat>: Adding Ship and Gameboard Class
Adding babel to be able to use jest inside of
webpack and creating a test to check that babel
and webpack are both working.

The ship class handles each boat with it's length,
hits and Sunk. alongside the class methods, hit()
(which adds a hit to the boat) and isSunk() (
this check to make sure that the boat is sunk or
not).

The Gameboard class handles the creation of a new
gameboard when creating a new gameboard, which is
a 10 x 10 2-d array to handle the boards data.
The method placeBoat() help to place a boat into
the 2-D array and makes sure the input is valid
and add the ship into the array index.
The method receiveShot() this handles the shots
onto the boats, so it checks to see if the boats
are their and if not it add 'missed' to the array
and also those co-ordinates to a missedShots array.
If the boat is at the index then runs the hit()
function.