<Feat/Fix>[gameboard]: refactor attack and add isBoatSunk
Refactoring the `receiveAttack` function to separate
the row and column into variables to make it more
readable.

Fixing the gameboard builder as it was referencing
just one array so it was copying each row of the
game to the corresponding rows.

Fixing the `missedShot` as it was adding 'MISS' to
the variable and not to the actual board so that
has been corrected.

Adding the `isBoatSunk` which check the board to
see if all the boats are sunk by running the
`Ship` function `isSunk` to make sure they are or
are not sunk, then return true if they are sunk
and false if any of the boats or Sunk = false.