<Feat/Fix>: apply GameFlow and fix object conditional
Adding a displayBoard() to display the preset boats
by using fillGame()which fills both players boards
with boats.

Creating GameFlow() to keep track of active players
and keep track of rounds and display onto the grid
for each player.
Need to fix the turn taking as it staying the same
player on every round.

Applied a table grid to keep track of the UI of the
battleship board for both players. each one ID is
their 2D array co-ordinates.

fixed the conditional for checking if the element
in an array is an object or null. tests have been
updated.

added basic styling for the tables and general
layout.

Pseudo is inside the README to follow the game flow.