<Feat/fix>: Added computer player and
Adding a computer player that selects random numbers
between 0 and 9 and then use them as co-ordinates
for the grid.

fixed activePlayer and each player now takes it in
turns to  play.

Feature of the non active player grid to disable
pointer events so there is no random clicking.

Feature of a dialog thats shows when one of the
players have won and shows who.

Feature of a restart button, works for the first round
but it then adds multiple eventlisteners on which
it runs the players go that many times, solution needs
to be only calling the eventlistener only on the first
load of the game.