<fix>: Changing GameFlow into a class.
Moving the game flow functions into a class to make
it easier to cycle through the game but also to
restart the game and initialize and new one game.

compressing the index.js into more readable chunks
and using the gameFlow class to make it so much
easier to start a new game and prevent any
duplication due to event listeners, making sure to
add event listener once and only once.