<feat>[randomPlaceBoat.js]: Adding randomize boat placement
Adding the feature of adding the boats randomly to
the grid, for both the real player and computer.
The change of placeBoat method to check the cells
in the grid first before placing the boat in there
to make sure there is not clashing of boats.

All in working order, just needs styling.